fastapi/pyproject.toml
[tool.poetry]
name = "repl-nix-agenticelectionpredictor"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.47.0"
uvicorn = "^0.30.6"
python-multipart = "^0.0.10"
fastapi = "^0.115.0"
psycopg2-binary = "^2.9.9"
pydantic = "^2.9.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
aiohttp = "^3.10.5"
pandas = "^2.2.3"
joblib = "^1.4.2"
scikit-learn = "^1.5.2"
sqlalchemy = "^2.0.35"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
flask = "^3.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
